Distribution/files/requirements.txt
flask
matplotlib
numpy
requests
pandas
loguru
psycopg2-binary
sqlalchemy
fake_useragent